Refactor: Pages 작성 규칙 변경 및 폴더 규직 변경 적용
- pages 규칙 변경
  - pages/**/*/index.tsx 는 최소한의 컴포넌트만 작성합니다.
  - 하나의 컴포넌트의 경우 components/views/*View.tsx의 이름을 가진 컴포넌트를 사용합니다.
- 폴더 규칙 변경
  - components/하위 *.tsx 파일의 경우 해당브랜치명의 폴더를 생성한 이후 내부에 기능단위로 구분을 하여 폴더를 생성 혹은 파일을 생성해서 작성합니다.
